refactor: navigation bar
