+++ src/components/Desktop/Overview/MainView.vue
@@ -1,14 +1,14 @@
 <script lang="ts" setup>
 import AppBar from '@/components/Desktop/AppBar.vue';
-import SideBar from '@/components/Desktop/SideBar.vue';
+import NavigationBar from '@/components/NavigationBar.vue';
 import Content from '@/components/Desktop/Overview/OverviewContent.vue';
 </script>
 
 <template>
-  <SideBar />
+  <NavigationBar />
   <div class="flex flex-col">
     <AppBar />
-    <content />
+    <Content />
   </div>
 </template>
