+++ src/components/Mobile/AppBar.vue
@@ -47,7 +47,7 @@
 
 .nav {
   position: fixed;
-  z-index: 9999;
+  z-index: auto;
   top: 0;
 
   width: 100%;
